Phase 20 chunk 213 (ADR-016): RawInt control words + conservative GC scan
Fixes the recursive-tabling heap-GC corruption and re-enables auto-collection.

Root cause: get_level stored the captured cut barrier (a choice-point stack
index) into a Y-slot as a plain Cell == Tag.Ref, indistinguishable from a heap
self-pointer. A GC scanning that slot relocated the barrier as a heap index, so
the next '!' cut to garbage (IndexOutOfRange in CompactTrails). Tabling failed
because it leans on cuts heavily.

Fix:
- Tag.RawInt (0xD) + Cell.RawInt: every env/choice-point control word and the
  get_level cut barrier is a distinctly-tagged leaf the collector never follows
  or relocates. Zero hot-path cost. ViewGen reads via .Payload; int slots via
  (int)Data.
- MarkRoots/RelocateRoots rewritten to a conservative full-stack scan, replacing
  the precise frame-liveness walk that under-counted roots in the tabling
  fixpoint's reused stack. Complete by construction; MarkReferents bounds-guards
  every payload so a stale ref over-retains at worst, never crashes.
- Removed the dump/dry-run/poison diagnostic scaffolding.

Auto-collection re-enabled: EngineConfig.GcThreshold defaults to 1<<18 (256K
cells); SHUMWAY_GC_THRESHOLD=N overrides at runtime (0 disables).

Validation: tabling/bagof/findall/listing correct under SHUMWAY_GC_STRESS=1;
full suite green with auto-GC on; 3 new cut-barrier/findall GC regression tests;
9 layout-assertion tests updated to read control slots as RawInt.

Measured (Blint, 78 KB source, same build): heap high-water 64 MB -> 8 MB (~8x),
same lint output, doubling-realloc churn eliminated.

+++ tests/Shumway.Tests.Core/EnvFrameTests.cs
@@ -32,9 +32,10 @@ public class EnvFrameTests
 
         Assert.Equal(0, engine.E);                 // newE = previous stackTop = 0
         Assert.Equal(5, engine.StackTop);          // 3 control (CE,CP,N) + 2 Y slots
-        Assert.Equal(-1L, engine.GetStack(0).Data); // CE = previous _e
-        Assert.Equal(100L, engine.GetStack(1).Data); // CP = previous _cp
-        Assert.Equal(2L, engine.GetStack(2).Data);  // N = permanent count (ADR-016)
+        // Control slots are RawInt-tagged (ADR-016); read with (int)Data.
+        Assert.Equal(-1, (int)engine.GetStack(0).Data); // CE = previous _e
+        Assert.Equal(100, (int)engine.GetStack(1).Data); // CP = previous _cp
+        Assert.Equal(2, (int)engine.GetStack(2).Data);  // N = permanent count (ADR-016)
 
         // Y slots are REFs to fresh heap unbounds.
         AssertUnboundHeapRef(engine, engine.GetStack(3));
@@ -76,12 +77,12 @@ public class EnvFrameTests
         Assert.Equal(9, engine.StackTop);
 
         // Second frame's CE points at the first frame; CP is the most-recent _cp.
-        Assert.Equal((long)firstFrame, engine.GetStack(secondFrame + Engine.EnvCeOffset).Data);
-        Assert.Equal(8L, engine.GetStack(secondFrame + Engine.EnvCpOffset).Data);
+        Assert.Equal(firstFrame, (int)engine.GetStack(secondFrame + Engine.EnvCeOffset).Data);
+        Assert.Equal(8, (int)engine.GetStack(secondFrame + Engine.EnvCpOffset).Data);
 
         // First frame's contents are unchanged by the second Allocate.
-        Assert.Equal(-1L, engine.GetStack(firstFrame + Engine.EnvCeOffset).Data);
-        Assert.Equal(7L, engine.GetStack(firstFrame + Engine.EnvCpOffset).Data);
+        Assert.Equal(-1, (int)engine.GetStack(firstFrame + Engine.EnvCeOffset).Data);
+        Assert.Equal(7, (int)engine.GetStack(firstFrame + Engine.EnvCpOffset).Data);
     }
 
     [Fact]
@@ -116,7 +117,7 @@ public class EnvFrameTests
         engine.Allocate(2);                         // forces growth
 
         // First frame's saved CP should still be readable.
-        Assert.Equal(99L, engine.GetStack(0 + Engine.EnvCpOffset).Data);
+        Assert.Equal(99, (int)engine.GetStack(0 + Engine.EnvCpOffset).Data);
     }
 
     [Fact]
